requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
opencv-python==4.8.1.78
ultralytics==8.0.200
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
PyYAML==6.0.1
psutil==5.9.5